backend/requirements.txt
# === CORE BACKEND FRAMEWORK ===
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2

# === LOCAL LLaMA MODEL & LLM INTEGRATION ===
llama-cpp-python==0.2.24
torch==2.7.0
transformers==4.36.0
huggingface-hub==0.19.4

# === LANGCHAIN & RAG COMPONENTS ===
langchain==0.0.350
langchain-community==0.0.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3






